+++ include/seqan3/alignment/matrix/detail/coordinate_matrix.hpp
@@ -20,7 +20,7 @@
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
-#include <seqan3/core/simd/view_iota_simd.hpp>
+#include <seqan3/utility/simd/views/iota_simd.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/utility/type_traits/lazy_conditional.hpp>
