+++ test/performance/core/simd/CMakeLists.txt
@@ -1,2 +1,1 @@
-seqan3_benchmark(view_iota_simd_benchmark.cpp)
 seqan3_benchmark(view_to_simd_chunk_benchmark.cpp)
