+++ test/performance/utility/simd/CMakeLists.txt
@@ -1,1 +1,3 @@
+add_subdirectories ()
+
 seqan3_benchmark(simd_algorithm_benchmark.cpp)
