+++ test/performance/utility/simd/views/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_benchmark(iota_simd_benchmark.cpp)
