+++ test/snippet/utility/simd/views/iota_simd.cpp
