+++ test/snippet/utility/simd/views/iota_simd_transform.cpp
