+++ test/unit/core/simd/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectories()
 
 seqan3_test(debug_stream_simd_test.cpp)
-seqan3_test(view_iota_simd_test.cpp)
 seqan3_test(view_to_simd_test.cpp)
