+++ test/unit/utility/simd/CMakeLists.txt
@@ -1,2 +1,4 @@
+add_subdirectories()
+
 seqan3_test (simd_algorithm_test.cpp)
 seqan3_test (simd_test.cpp)
