+++ test/unit/utility/simd/views/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_test (iota_simd_test.cpp)
