+++ test/unit/utility/simd/views/iota_simd_test.cpp
@@ -10,11 +10,11 @@
 #include <vector>
 
 #include <seqan3/utility/simd/simd.hpp>
-#include <seqan3/core/simd/view_iota_simd.hpp>
+#include <seqan3/utility/simd/views/iota_simd.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/test/simd_utility.hpp>
 
-#include "../../range/iterator_test_template.hpp"
+#include "../../../range/iterator_test_template.hpp"
 
 template <typename simd_t>
 struct iota_simd_test : public ::testing::Test
